`BenchmarkHash` function, the parameter `h` is changed from `maphash.Hash` to `*maphash.Hash`
In the `benchmarkHash` function, the parameter `h` is changed from `maphash.Hash` to `*maphash.Hash`. This is because `maphash.Hash` is a struct, not an interface, and should be passed as a pointer.

In the `BenchmarkMaphash` function, when calling `benchmarkHash`, the `&` operator is added before `h` to pass the address of `h` instead of the value.